requirements.txt
aerich==0.7.1
aiofiles==0.8.0
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
aiosmtplib==1.1.6
aiosqlite==0.17.0
anyio==3.7.0
async-timeout==4.0.2
asyncmy==0.2.8
attrs==23.1.0
bcrypt==3.2.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==7.1.2
cryptography==39.0.1
dictdiffer==0.9.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.1.3
exceptiongroup==1.1.1
fast-captcha==0.1.3
fastapi==0.95.2
fastapi-pagination==0.9.1
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
httptools==0.1.2
idna==3.4
iso8601==1.1.0
loguru==0.6.0
multidict==6.0.4
openai==0.27.8
passlib==1.7.4
path==15.1.2
Pillow==8.4.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.5
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
supervisor==4.2.5
tiktoken==0.4.0
tomlkit==0.11.8
tortoise-orm==0.19.3
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.13.4
uvloop==0.17.0
wait-for-it==2.2.1
watchgod==0.8.2
websockets==8.1
yarl==1.9.2
